Create BatchScript2
This is updated batch file in which problem for .mtl file (materials could not be loaded) is resolved. It is done in a way that resources folder is removed and new unique directory is created (using BUILD NUMBER environment variable) in which new resource folder is placed. Files are unzipped in the new resource folder then unity project is run once so that unity detects new unique directory and update its materials by creating Material folders automatically. As materials has been generated so we do not need .mtl files (or to convert them to .txt). Now assets are named with their corresponding folder as their asset bundles along with materials created. So that now exported asset bundles also contain materials. It can be tested in unity project that asset loaded from asset bundle also contain materials and textures.